+++ Sources/Kommunicate/Classes/KMConversationViewConfiguration.swift
@@ -18,6 +18,8 @@ public struct KMConversationViewConfiguration {
     public var startNewButtonIcon: UIImage? = UIImage(named: "icon_new_chat_red", in: Bundle.kommunicate, compatibleWith: nil)
     /// If enabled, the user can't send a message when a conversation is assigned to a bot.
     public var restrictMessageTypingWithBots = false
+    public var toolbarSubtitleText = ""
+    public var toolbarSubtitleRating: Float = -1.0
 
     public init() {}
 }